SONAR-5471 drop migrations related to column rules.characteristic_id
