+++ sonar-db/src/test/resources/org/sonar/db/version/v451/AddMissingCustomRuleParametersMigrationTest/execute-result.xml
@@ -5,7 +5,6 @@
          description="Architectural constraint" status="READY" priority="1" language="xoo"
          note_data="[null]" note_user_login="[null]" note_created_at="[null]" note_updated_at="[null]"
          description_format="HTML" tags="[null]" system_tags="[null]" plugin_config_key="[null]"
-         characteristic_id="[null]" default_characteristic_id="[null]"
          remediation_function="[null]" default_remediation_function="[null]"
          remediation_coeff="[null]" default_remediation_coeff="[null]"
          remediation_offset="[null]" default_remediation_offset="[null]"
@@ -21,7 +20,6 @@
          description="Architectural constraint 2" status="READY" priority="1" language="xoo"
          note_data="[null]" note_user_login="[null]" note_created_at="[null]" note_updated_at="[null]"
          description_format="HTML" tags="[null]" system_tags="[null]" plugin_config_key="[null]"
-         characteristic_id="[null]" default_characteristic_id="[null]"
          remediation_function="[null]" default_remediation_function="[null]"
          remediation_coeff="[null]" default_remediation_coeff="[null]"
          remediation_offset="[null]" default_remediation_offset="[null]"
