+++ sonar-db/src/test/resources/org/sonar/db/version/v55/FeedRulesLongDateColumnsTest/schema.sql
@@ -15,8 +15,6 @@ CREATE TABLE "RULES" (
   "NOTE_USER_LOGIN" VARCHAR(255),
   "NOTE_CREATED_AT" TIMESTAMP,
   "NOTE_UPDATED_AT" TIMESTAMP,
-  "CHARACTERISTIC_ID" INTEGER,
-  "DEFAULT_CHARACTERISTIC_ID" INTEGER,
   "REMEDIATION_FUNCTION" VARCHAR(20),
   "DEFAULT_REMEDIATION_FUNCTION" VARCHAR(20),
   "REMEDIATION_COEFF" VARCHAR(20),
